associate tokens to facade
