+++ packages/hedera-wallet-snap/packages/snap/src/index.ts
@@ -27,7 +27,6 @@ import { transferCrypto } from './rpc/transactions/transferCrypto';
 import { SnapAccounts } from './snap/SnapAccounts';
 import { SnapState } from './snap/SnapState';
 import { WalletSnapParams } from './types/state';
-import { associateTokens } from './rpc/hts/associateTokens';
 import { createToken } from './rpc/hts/createToken';
 import { SignMessageCommand } from './commands/SignMessageCommand';
 import { HederaUtils } from './utils/HederaUtils';
@@ -39,6 +38,7 @@ import { StakeHbarFacade } from './Facades/StakeHbarFacade';
 import { ApproveAllowanceFacade } from './Facades/ApproveAllowanceFacade';
 import { DeleteAllowanceFacade } from './Facades/DeleteAllowanceFacade';
 import { DeleteAccountFacade } from './Facades/DeleteAccountFacade';
+import {AssociateTokensFacade} from "./Facades/AssociateTokensFacade";
 
 /**
  * Handle incoming JSON-RPC requests, sent through `wallet_invokeSnap`.
@@ -217,7 +217,7 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
       HederaUtils.isValidAssociateTokensParams(request.params);
       return {
         currentAccount: state.currentAccount,
-        receipt: await associateTokens(walletSnapParams, request.params),
+        receipt: await AssociateTokensFacade.associateTokens(walletSnapParams, request.params),
       };
     }
     case 'hts/createToken': {
